Extract TargetRuby class
