+++ lib/rubocop.rb
@@ -616,4 +616,5 @@ require_relative 'rubocop/cli/command/show_cops'
 require_relative 'rubocop/cli/command/version'
 require_relative 'rubocop/options'
 require_relative 'rubocop/remote_config'
+require_relative 'rubocop/target_ruby'
 require_relative 'rubocop/yaml_duplication_checker'
