+++ spec/rubocop/cop/style/redundant_freeze_spec.rb
@@ -54,7 +54,7 @@ RSpec.describe RuboCop::Cop::Style::RedundantFreeze do
   end
 
   context 'when the receiver is a frozen string literal' do
-    if RuboCop::ConfigValidator::KNOWN_RUBIES.include?(3.0)
+    if RuboCop::TargetRuby.supported_versions.include?(3.0)
       context 'when the target ruby version >= 3.0' do
         let(:ruby_version) { 3.0 }
